requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> python=3,10
# platform: Ubuntu-22.04
absl-py=2.2.2=pypi_0
anyio=4.9.0=pypi_0
asttokens=3.0.0=pypi_0
astunparse=1.6.3=pypi_0
attrs=25.3.0=pypi_0
biopython=1.85=pypi_0
biotite=1.2.0=pypi_0
biotraj=1.2.2=pypi_0
brotli=1.1.0=pypi_0
causal-conv1d=1.5.0.post8=pypi_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
cloudpathlib=0.21.1=pypi_0
decorator=5.2.1=pypi_0
einops=0.8.1=pypi_0
esm=3.2.1=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.0=pypi_0
filelock=3.18.0=pypi_0
flatbuffers=25.2.10=pypi_0
fsspec=2025.3.2=pypi_0
gast=0.6.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.71.0=pypi_0
h11=0.16.0=pypi_0
h5py=3.13.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.31.2=pypi_0
hyperpyyaml=1.2.2=pypi_0
idna=3.10=pypi_0
imbalanced-learn=0.13.0=pypi_0
imblearn=0.0=pypi_0
ipython=8.37.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.0=pypi_0
keras=3.9.2=pypi_0
libclang=18.1.1=pypi_0
mamba-ssm=2.2.4=pypi_0
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.5.1=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.1=pypi_0
msgpack-numpy=0.4.8=pypi_0
namex=0.0.9=pypi_0
networkx=3.4.2=pypi_0
ninja=1.11.1.4=pypi_0
numpy=2.1.3=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
opt-einsum=3.4.0=pypi_0
optree=0.15.0=pypi_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.2.1=pypi_0
pip=25.1=pypi_0
prompt-toolkit=3.0.51=pypi_0
protobuf=5.29.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pygments=2.19.1=pypi_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-metric-learning=2.8.1=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=14.0.0=pypi_0
ruamel-yaml=0.18.12=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.3=pypi_0
sentencepiece=0.2.0=pypi_0
seqtk=0.5.0=pypi_0
setuptools=78.1.1=pypi_0
six=1.17.0=pypi_0
sklearn-compat=0.1.3=pypi_0
sniffio=1.3.1=pypi_0
speechbrain=1.0.3=pypi_0
stack-data=0.6.3=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.1.2=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tokenizers=0.21.1=pypi_0
torch=2.6.0=pypi_0
torchaudio=2.6.0=pypi_0
torchtext=0.18.0=pypi_0
torchvision=0.21.0=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.48.1=pypi_0
triton=3.2.0=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pypi_0
wrapt=1.17.2=pypi_0
zstd=1.5.7.2=pypi_0